Major commit!
GatingSet query modified to include gating set description column to be used as a tool tip in the combo box, for which this query is used. Also had to include gsid column into the query in order for the LABKEY.ext.store to properly fetch the data generated by the query.

Population query modified to include the analysis column, so that proper filtering of the population set can be done. (not yet tested)

StudyVars query created to contain the set of associated study variables for each of the generated gating sets (analyses).

Plot.r :
Parameter added for the usage of hexbin - smoothing of the data for display, for large data sets can result in faster rendering times.
Now the table of gating sets is used, the path of the selected gating set/analysis is passed from JS so that R can read it from disk.

begin.html :
Modified to not load the module contents if the StudyVars query is not present in the opencyto_preprocessing schema, meaning that the OpenCytoVisualization module has been disabled (but its interface, the web part has not been removed from the home dashboard and still rendered by the server). Once issue #18 is resolved on the server side, this would need to be removed.

ClearableComboBox.js control added to allow for a one-button approach to clearing a combo box value.

OpenCytoVisualization.js :
All of the common functionality needed for the entire suite moved to OpenCyto.js.
strStudyVarName modified to include the Display field, the Value field and the Analysis field (the latter for filtering). The source for the strStudyVarName's data now comes from a custom table containing selected study variables for the selected analysis: the additional fields are now heavily used in order to speed up the computation time.
Added a slider for selecting the hexbin parameter value.
Fixed the generated graphic, so that if 2 identical web parts are present, each of the Resizable widgets is attached to the corresponding graphic.
Moved the analysis selection combo box to the first tab.
Replaced the 'drag and drop' instruction label with a quick tip, when hovering over the corresponding panel's header.
Implemented the logic so that once a value is selected from either x-axis or y-axis combo box, that value is not present in the other combo box's choices (close issue #14).

Overall streamlined and improved the layout design.